Refactorings to begin getting rid of rustc_codegen_utils
