+++ src/librustc_session/lib.rs
@@ -21,3 +21,5 @@ pub mod search_paths;
 
 mod session;
 pub use session::*;
+
+pub mod output;
